pyproject.toml
[project]
name = "TickTick-Widget"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "click>=8.2.1",
    "loguru>=0.7.3",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
    "pyticktick>=0.1.0",
]
